v1.1111

+++ src/main/java/org/edu/infi_payment_system/User/service/UserServiceImpl.java
@@ -5,8 +5,12 @@ import lombok.RequiredArgsConstructor;
 import org.edu.infi_payment_system.User.dto.request.LoginRequestDto;
 import org.edu.infi_payment_system.User.dto.request.RegisterRequestDto;
 import org.edu.infi_payment_system.User.dto.response.UserResponseDto;
+import org.edu.infi_payment_system.User.entity.User;
+import org.edu.infi_payment_system.User.exception.InvalidCredentialsException;
+import org.edu.infi_payment_system.User.exception.UserAlreadyExistsException;
 import org.edu.infi_payment_system.User.mapper.UserMapper;
 import org.edu.infi_payment_system.User.repository.UserRepository;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -21,27 +25,53 @@ public class UserServiceImpl implements UserService{
     private final UserMapper userMapper;
     private final PasswordEncoder passwordEncoder;
 
+
     @Override
     public UserResponseDto registerUser(@Valid RegisterRequestDto requestDto){
-        return new UserResponseDto();
+
+        if(userRepository.existsByEmail(requestDto.getEmail())){
+            throw new UserAlreadyExistsException("User already exists with this email");
+        }
+
+        User user = userMapper.toEntity(requestDto);
+        user.setPassword(passwordEncoder.encode(requestDto.getPassword()));
+
+        User savedUser = userRepository.save(user);
+        return userMapper.toResponseDto(savedUser);
     }
 
     @Override
     public String loginUser(@Valid LoginRequestDto requestDto){
+
+        User user = userRepository.findByEmail(requestDto.getEmail())
+                .orElseThrow(() -> new UsernameNotFoundException("Invalid email"));
+
+        if(!passwordEncoder.matches(requestDto.getPassword(), user.getPassword())){
+            throw new InvalidCredentialsException("Invalid password");
+        }
+
         return "dummy token - H389Y3BNDO3289NMNNDU3820BJD62J";
     }
     @Override
     public UserResponseDto getUserById(Long id){
-        return new UserResponseDto();
+
+        User user = userRepository.findById(id).orElseThrow( () -> new UsernameNotFoundException("user not found with id "+ id));
+
+        return userMapper.toResponseDto(user);
     }
 
     @Override
     public List<UserResponseDto> getAllUsers(){
-        return userRepository.findAll().stream().map(user-> new UserResponseDto()).collect(Collectors.toList());
+        return userRepository.findAll().stream()
+                .map(userMapper::toResponseDto)
+                .collect(Collectors.toList());
     }
 
     @Override
     public void deleteUser(Long id){
+        if(!userRepository.existsById(id)){
+            throw new UsernameNotFoundException("User not found with id " + id);
+        }
         userRepository.deleteById(id);
     }
 }
